+++ src/software/sensor_fusion/filter/ball_filter.cpp
@@ -190,8 +190,8 @@ std::optional<size_t> BallFilter::getAdjustedBufferSize(
     double linear_scaling_factor = linear(min_max_magnitude_average, linear_offset,
                                           buffer_size_velocity_magnitude_diff);
     int buffer_size =
-        max_buffer_size -
-        static_cast<unsigned int>(std::floor(linear_scaling_factor * buffer_size_diff));
+        static_cast<int>(max_buffer_size) -
+        static_cast<int>(std::floor(linear_scaling_factor * buffer_size_diff));
 
     return static_cast<size_t>(buffer_size);
 }
