+++ src/software/world/robot.cpp
@@ -29,7 +29,7 @@ Robot::Robot(RobotId id, const RobotState &initial_state, const Timestamp &times
 }
 
 Robot::Robot(const TbotsProto::Robot &robot_proto)
-    : id_(robot_proto.id()),
+    : id_(static_cast<int>(robot_proto.id())),
       current_state_(RobotState(robot_proto.current_state())),
       timestamp_(Timestamp::fromTimestampProto(robot_proto.timestamp())),
       robot_constants_(DEFAULT_ROBOT_CONSTANTS)
